+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to the "P2 Land Review Feature" extension will be documented
 
 ## [Unreleased]
 -0.0.3
-  - fix a bug related to verify 'Use And'
   - add validation to find duplicate test case id
+  - fix a bug related to verify 'Use And'
+  - fix a bug related to detecting 'Use Scenario with Examples'
 - 0.0.2
   - re-written to use Gherkin parser
   - implement more validation rules